backend/requirements.txt
# Core FastAPI
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.5
python-dotenv==1.2.2

# MongoDB
motor==3.3.1
pymongo==4.5.0

# Calendar sync
icalendar==5.0.13
requests==2.33.1

# Scheduler
APScheduler==3.10.4

# CORS
starlette==0.37.2
resend==2.28.1
